+++ R/plot.snowprofileSet.R
@@ -2,19 +2,25 @@
 #'
 #' A flexible function to plot multiple snowprofiles either in a timeseries or various types of groups.
 #'
-#' The routine allows you to plot colored sequences only, or to include hardness profile information as well. See parameter
+#' The routine allows you to plot coloured sequences only, or to include hardness profile information as well. See parameter
 #' `hardnessResidual` and the examples for more details.
 #' To change the font size of labels etc, use `par()` with the parameters `cex.lab`, `cex.axis`, etc.
 #'
 #' @param x An object of class [snowprofileSet]
 #' @param SortMethod How to arrange profiles along the x-axis. Options include timeseries (default = 'time'), in existing order of Profiles list ('unsorted'), sorted by HS ('hs'), or elevation ('elev')
-#' @param ColParam What parameter to show with colour. So far the following types are available: graintype (default), hardness, temperature, density, grainsize.
+#' @param ColParam What parameter to show with colour. So far the following types are available: "gtype", "hardness", "density", "temp", "gsize", "ssi", "p_unstable", "crit_cut_length", "rta", "percentage".
 #' @param TopDown Option to plot by depth instead of height with zero depth on top of plot (default = FALSE)
 #' @param DateStart Start date for timeseries plots (`SortMethod = 'time'`). If not provided, the function takes the date range from Profiles (default = NA).
 #' @param DateEnd End date for timeseries plots (`SortMethod = 'time'`). If not provided, the function takes the date range from Profiles (default = NA).
 #' @param Timeseries_labels Label Saturdays "weekly", "monthly", or `NA`
 #' @param ylim Vertical range of plot
 #' @param OutlineLyrs Switch for outlining layers (default = FALSE)
+#' @param emphasizeLayers emphasize layers with different transparency than others, or a different color altogether? then set this argument to `TRUE` if you want to emphasize all
+#' labeled layers of interest (aka weak layers), or provide a named list with arguments to a function call to [findPWL] to define which layers to emphasize.
+#' Set either `colAlpha` or `colEmphasis` to make the emphasis apparent.
+#' @param colAlpha the transparency setting for all layers (except the ones to be emphasized if you want to emphasize any). This can be useful for example
+#' if you want to overplot the grain type sequences with another variable, e.g. a percentage from a distribution.
+#' @param colEmphasis the color of the layers to be emphasized (only if you want a different color than defined by `ColParam`)
 #' @param OutlineProfile vector of profile indices that will be outlined to highlight them
 #' @param HorizGrid Draw horizontal grid at layer heights (default = TRUE)
 #' @param yaxis draw a y-axis? (either `FALSE`, `TRUE` draws yaxis left, `"right"` draws yaxis on the right plot side)
@@ -28,7 +34,8 @@
 #' @param yPadding Padding between ylim and limits of data, default = 10.
 #' Note that R will still put padding by default. If you want to prohibit that entirely, specify `xaxs ='i'`, or `yaxs = 'i'`.
 #' @param xPadding Padding between xlim and limits of data, default = 0.5.
-#' Note that R will still put padding by default. If you want to prohibit that entirely, specify `xaxs = 'i'`, or `yaxs = 'i'`.
+#' Note that R will still put padding by default. If you want to prohibit that entirely, specify `xaxs = 'i'`, or `yaxs = 'i'`. For xPadding, you can provide either a scalar, or
+#' a length 2 numeric for left and right hand side, respectively.
 #' @param hardnessResidual Value within `[0, 1]` to control the minimum horizontal space of each layer that will be colored
 #' irrespective of the layer's hardness. A value of `1` corresponds to no hardness being shown.
 #' @param hardnessScale A scaling factor that exaggerates the hardness profile to subsequent cells on the x-axis. Useful
@@ -37,6 +44,7 @@
 #' @param hardnessOffset offsets the profile location on the x-axis
 #' @param k a sorting vector if `SortMethod = "presorted"`.
 #' @param offset Provide a Date or POSIXct offset if you want to offset the vertical snow height/depth axis so that the offset date aligns with snow depth/height 0.
+#' @param add add the plot to an existing plot, or create new plot?
 #' @param ... Additional parameters passed to plot()
 #'
 #' @seealso [plot.snowprofile], [SPgroup]
@@ -84,18 +92,31 @@
 #'      offset = as.Date("2019-01-21"), xlab = "<-- Jan 21 PWL exists | does not exist -->")
 #' abline(v = max(which(pwl_exists[k]))+0.5, lty = "dashed")
 #'
+#' ## Emphasize specific layers
+#' ## (i) all labeled layers of interest:
+#' SPgroup <- snowprofileSet(lapply(SPgroup, labelPWL))  # label layers with default settings
+#' plot(SPgroup, SortMethod = "hs", emphasizeLayers = TRUE, colAlpha = 0.3)
+#' ## (ii) specific individual layers:
+#' plot(SPgroup, SortMethod = "hs",
+#'      emphasizeLayers = list(pwl_gtype = c("SH", "DH"), pwl_date = "2019-01-21"),
+#'      colAlpha = 0.3, colEmphasis = "black")
+#'
+#'
 #' @export
 #'
 
 plot.snowprofileSet <- function(x,
                                 SortMethod = c("time", "unsorted", "hs", "elev", "presorted"),
-                                ColParam = c("gtype", "hardness", "temperature", "density", "gsize"),
+                                ColParam = c("gtype", "hardness", "density", "temp", "gsize", "ssi", "p_unstable", "crit_cut_length", "rta", "percentage"),
                                 TopDown = FALSE,
                                 DateStart = NA,
                                 DateEnd = NA,
                                 Timeseries_labels = c("weekly", "monthly", NA),
                                 ylim = NULL,
                                 OutlineLyrs = FALSE,
+                                emphasizeLayers = NULL,
+                                colAlpha = NA,
+                                colEmphasis = NA,
                                 OutlineProfile = NULL,
                                 HorizGrid = TRUE,
                                 yaxis = TRUE,
@@ -112,6 +133,7 @@ plot.snowprofileSet <- function(x,
                                 hardnessOffset = -0.5,
                                 k = NULL,
                                 offset = as.Date(NA),
+                                add = FALSE,
                                 ...) {
 
 
@@ -127,14 +149,14 @@ plot.snowprofileSet <- function(x,
     stop(paste("SortMethod", SortMethod, "not supported ('time', 'hs', 'elev', 'unsorted')"))
 
   ## ColParam
-  if (!(ColParam[1] %in% c("gtype", "hardness", "density", "temp", "gsize", "ssi")))
-    stop(paste("ColParam", ColParam, "not supportd"))
+  if (!(ColParam[1] %in% c("gtype", "hardness", "density", "temp", "gsize", "ssi", "p_unstable", "crit_cut_length", "rta", "percentage")))
+    stop(paste("ColParam", ColParam, "not supported"))
 
 
   ## ---- Filter and sort profiles ----
 
   ## Calculate metadata
-  Metadata <- summary(Profiles)
+  Metadata <- summary(Profiles, fast = FALSE)
 
   ## Filter dates
   if (SortMethod[1] == "time") {
@@ -169,8 +191,19 @@ plot.snowprofileSet <- function(x,
 
   ## offset layer locations based on a given offset Date
   if (!is.na(offset)) {
-    if (!(inherits(offset, "Date") | inherits(offset, "POSIXct"))) stop("offset needs to be of class Date or POSIXct!")
-    offset <- as.POSIXct(offset)
+    # tz_used <- NA
+    # try({tz_used <- attr(Profiles[[1]]$layers$datetag[1], 'tz')})
+    # try({tz_used <- attr(Profiles[[1]]$layers$ddate[1], 'tz')})
+    # if (inherits(offset, "Date")) {
+    #   if (is.na(tz_used)) {
+    #     offset <- as.POSIXct(offset)
+    #   } else {
+    #     offset <- as.POSIXct(offset, tz = tz_used)
+    #   }
+    # } else if (!inherits(offset, "POSIXct")) stop("offset needs to be of class Date or POSIXct!")
+    if (inherits(offset, "POSIXct")) {
+      offset <- as.Date(as.character(offset))
+    } else if (!inherits(offset, "Date")) stop("offset needs to be of class Date or POSIXct!")
 
     offset_profiles <- lapply(Profiles, function(sp) {
       sp_d <- deriveDatetag(sp)
@@ -179,6 +212,21 @@ plot.snowprofileSet <- function(x,
       return(sp_d)
     })
     Profiles <- snowprofileSet(offset_profiles)
+  }
+
+  ## compute which layers to emphasize (if desired)
+  if (!all(is.null(emphasizeLayers))) {
+    emphasizeLayersofinterest <- FALSE
+    if (isTRUE(emphasizeLayers)) {
+      emphasizeLayersofinterest <- TRUE
+    } else if (is.list(emphasizeLayers)) {
+      Profiles <- snowprofileSet(lapply(Profiles, function(sp) {
+        sp$layers$toEmphasize <- FALSE
+        sp$layers$toEmphasize[do.call("findPWL", c(quote(sp), emphasizeLayers))] <- TRUE
+        sp
+      }))
+    }
+    if (is.na(colAlpha) & is.na(colEmphasis)) warning("If you want to emphasizeLayers, you also need to specify a colAlpha or colEmphasis to make the emphasis apparent!")
   }
 
   ## Rbind all the profiles to get table with all layers
@@ -221,18 +269,37 @@ plot.snowprofileSet <- function(x,
     Lyrs$Col <- getColoursSnowTemp(Lyrs$temperature)
   } else if (ColParam[1] == "gsize") {
     Lyrs$Col <- getColoursGrainSize(Lyrs$gsize)
+  } else if (ColParam[1] == "p_unstable") {
+    Lyrs$Col <- getColoursStability(Lyrs$p_unstable, StabilityIndexThreshold = 0.77, StabilityIndexRange = c(0, 1))
+  } else if (ColParam[1] == "crit_cut_length") {
+    Lyrs$Col <- getColoursStability(Lyrs$crit_cut_length, StabilityIndexThreshold = 0.4, StabilityIndexRange = c(0, 3), invers = TRUE)
+  } else if (ColParam[1] == "rta") {
+    Lyrs$Col <- getColoursStability(Lyrs$rta, StabilityIndexThreshold = 0.8, StabilityIndexRange = c(0, 1))
+  } else if (ColParam[1] == "percentage") {
+    Lyrs$Col <- getColoursPercentage(Lyrs$percentage, ClrRamp = "Greys_transparent")
   } else {
     Lyrs$Col <- "#E8E8E8"
   }
 
+  if (!is.na(colAlpha) | !is.na(colEmphasis)) {
+    if (!all(is.null(emphasizeLayers))) {
+      if (!is.na(colEmphasis)) {
+        Lyrs$emphasisCol <- colEmphasis
+      } else {
+        Lyrs$emphasisCol <- Lyrs$Col
+      }
+    }
+    if (!is.na(colAlpha)) Lyrs$Col <- sapply(Lyrs$Col, function(cl) adjustcolor(cl, alpha.f = colAlpha))
+  }
 
   ## ---- Set plot dimensions ----
 
   ## xlim
-  xlim <- c(0.5 - xPadding, length(Profiles) + 0.5 + xPadding)
-  if (SortMethod[1] == "time")
-    xlim <- c(min(xx) - 1, max(xx + 1))
+  if (length(xPadding) == 2) xlim <- c(0.5 - xPadding[1], length(Profiles) + 0.5 + xPadding[2]) else xlim <- c(0.5 - xPadding, length(Profiles) + 0.5 + xPadding)
+  if (SortMethod[1] == "time") {
+    if (length(xPadding) == 2) xlim <- c(min(xx) - xPadding[1], max(xx + xPadding[2])) else xlim <- c(min(xx) - xPadding, max(xx + xPadding))
     if (any(is.na(xlim))) stop("Profiles do not contain date information, SortMethod 'time' is not available, try SorthMethod = 'unsorted'")
+  }
 
   ## ylim
   if (is.null(ylim)) {
@@ -248,17 +315,20 @@ plot.snowprofileSet <- function(x,
       ymax <- ifelse(TopDown, 0, getMaxHSequivalent(Metadata) + yPadding)
       ylim <- c(ymin, ymax)
     }
-  }
+  } else if (is.call(ylim)) ylim <- eval(ylim)  # allows users to send quoted ylims, e.g. quote(c(max(min(y1), -100), min(100, max(y2))))
 
 
   ## ---- Draw plot ----
 
   ## Initialize empty plot
-  plot(NA, NA, type = "n",
-       xlim = xlim, ylim = ylim,
-       xaxt = "n", yaxt = "n", axes = FALSE,
-       xlab = "", ylab = "",
-       ...)
+  if (!add) {
+    plot(NA, NA, type = "n",
+         xlim = xlim, ylim = ylim,
+         xaxt = "n", yaxt = "n", axes = FALSE,
+         xlab = "", ylab = "",
+         ...)
+  }
+
 
   if (exists("offset_profiles")) abline(h = 0, lwd = 1.5)
 
@@ -283,6 +353,25 @@ plot.snowprofileSet <- function(x,
          ytop = sapply(OutlineProfile, function(op) max(y2[xx %in% op])),
          col = NA,
          border = "#606060")
+  }
+
+  ## Emphasize specific, individual layers
+  if (!all(is.null(emphasizeLayers))) {
+    if (emphasizeLayersofinterest) {
+      rect(xleft = xx[Lyrs$layerOfInterest] + hardnessOffset,
+           ybottom = y1[Lyrs$layerOfInterest],
+           xright = xx[Lyrs$layerOfInterest] + hardnessOffset + hardnessResidual + Hardness[Lyrs$layerOfInterest]*hardnessScale*(1-hardnessResidual),
+           ytop = y2[Lyrs$layerOfInterest],
+           col = Lyrs$emphasisCol[Lyrs$layerOfInterest],
+           border = ifelse(OutlineLyrs == FALSE, NA, "#606060"))
+    } else if ("toEmphasize" %in% names(Lyrs)) {
+      rect(xleft = xx[Lyrs$toEmphasize] + hardnessOffset,
+           ybottom = y1[Lyrs$toEmphasize],
+           xright = xx[Lyrs$toEmphasize] + hardnessOffset + hardnessResidual + Hardness[Lyrs$toEmphasize]*hardnessScale*(1-hardnessResidual),
+           ytop = y2[Lyrs$toEmphasize],
+           col = Lyrs$emphasisCol[Lyrs$toEmphasize],
+           border = ifelse(OutlineLyrs == FALSE, NA, "#606060"))
+    }
   }