+++ R/summary.snowprofileSet.R
@@ -3,6 +3,7 @@
 #' Wrapper for [summary.snowprofile], which only returns metadata for a single snowprofile object. summary.snowprofileSet provides metadata for multiple snowprofiles, which is useful for subsetting.
 #'
 #' @param object list of snowprofile objects
+#' @param fast boolean switch to speed up computations, see [summary.snowprofile]
 #' @param ... additional arguments for generic method
 #'
 #' @return data.frame
@@ -24,14 +25,21 @@
 #' summary(Shallow)
 #' Week2 <- SPtimeline[summary(SPtimeline)$date > '2017-12-15']
 #'
+#'
+#' ## time comparison of fast--slow implementation
+#' ## expect 20 sec runtime
+#' # rbenchmark::benchmark(fast = {Metadata <- summary(SPgroup, fast = TRUE)},
+#' #                       slow = {Metadata <- summary(SPgroup, fast = FALSE)},
+#' #                       replications = 10**3)
+#'
 #' @export
 #'
 #' @import data.table
 #'
-summary.snowprofileSet <- function(object, ...) {
+summary.snowprofileSet <- function(object, fast = TRUE, ...) {
 
   ## Produce list of summaries of individual profiles
-  Summaries <- lapply(object, summary)
+  Summaries <- lapply(object, summary, fast = fast)
   Summaries <- data.table::rbindlist(Summaries, fill = TRUE)
   Summaries <- as.data.frame(Summaries)